+++ common/src/main/java/com/teamresourceful/resourcefulconfig/client/components/options/OptionItem.java
@@ -11,7 +11,12 @@ import net.minecraft.client.gui.components.MultiLineTextWidget;
 import net.minecraft.client.gui.components.StringWidget;
 import net.minecraft.client.gui.layouts.EqualSpacingLayout;
 import net.minecraft.client.gui.layouts.LinearLayout;
+import net.minecraft.client.gui.screens.ConfirmLinkScreen;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.network.chat.ClickEvent;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.Style;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
@@ -61,6 +66,7 @@ public class OptionItem extends ContainerWidget implements ListWidget.Item {
                 new MultiLineTextWidget(this.description, font)
                         .setCentered(false)
                         .setMaxWidth(half)
+                        .configureStyleHandling(true, OptionItem::handleStyle)
         );
 
         LinearLayout options = LinearLayout
@@ -99,5 +105,21 @@ public class OptionItem extends ContainerWidget implements ListWidget.Item {
         this.setWidth(width);
         if (!changed) return;
         init();
+    }
+
+    private static void handleStyle(@NotNull Style style) {
+        var event = style.getClickEvent();
+        var mc = Minecraft.getInstance();
+        if (event != null) {
+            switch (event) {
+                case ClickEvent.CopyToClipboard clipboard -> mc.keyboardHandler.setClipboard(clipboard.value());
+                case ClickEvent.OpenUrl link -> {
+                    Screen screen = Minecraft.getInstance().screen;
+                    if (screen == null) return;
+                    ConfirmLinkScreen.confirmLinkNow(screen, link.uri());
+                }
+                default -> {}
+            }
+        }
     }
 }